requirements.txt
boto3~=1.28.2
botocore~=1.31.2
psycopg2~=2.9.6
snowflake-connector-python
pandas~=1.5.3